+++ upload/library/Iversia/FAQ/AlertHandler/Question.php
@@ -5,6 +5,7 @@ class Iversia_FAQ_AlertHandler_Question extends XenForo_AlertHandler_Abstract
     public function getContentByIds(array $contentIds, $model, $userId, array $viewingUser)
     {
         // alert_{contentType}_like
+        // alert_{contentType}_answered
         return $model->getModelFromCache('Iversia_FAQ_Model_Question')->getQuestionsByIds($contentIds);
     }
 }